Extract `balances$` into `clients/balances`
- Remove few Balance* types from `wallet/types` we already have in
`client/types`
