+++ src/renderer/components/wallet/assets/AssetsTableCollapsable.tsx
@@ -11,8 +11,9 @@ import * as O from 'fp-ts/lib/Option'
 import { useIntl } from 'react-intl'
 
 import { getPoolPriceValue } from '../../../services/binance/utils'
+import { BalancesRD } from '../../../services/clients'
 import { PoolDetails } from '../../../services/midgard/types'
-import { ChainBalances, BalancesRD, ApiError, ChainBalance } from '../../../services/wallet/types'
+import { ChainBalances, ApiError, ChainBalance } from '../../../services/wallet/types'
 import { PricePool } from '../../../views/pools/Pools.types'
 import { ErrorView } from '../../shared/error/'
 import { AssetIcon } from '../../uielements/assets/assetIcon'
